just need to fix dirty

+++ src/Block.java
@@ -3,7 +3,7 @@ public class Block {
     boolean valid=false;
     boolean dirty;
     int entry_time;
-    int LRU;
+    long LRU;
     String block_index;
     int set_index;
     String tag;
